+++ changelog.md
@@ -16,6 +16,9 @@
   - Expanded [client.writeClipboard](docs/scripting.md) added html argument to add formated html as well as plain text
   - Expanded [client.writeClipboardHTML](docs/scripting.md) write HTML markup to clipboard to produce formated text for applications that support HTML pasting
   - Expanded [client.readClipboardHTML](docs/scripting.md) read HTML markup from clipboard if clipboard has HTML support
+  - Added total time to [#testspeed](docs/commands.md) command
+  - Added [#testspeedfile](docs/commands.md) to load a file and run timed test
+  - Added [#testfile](docs/commands.md) to load a file
 - **Changed:**
   - Update electron from 3.0.5 to 3.0.6
   - Log viewer uses simpler HTML generation
@@ -23,7 +26,7 @@
   - Status: Fix sleet weather overlay for all themes
   - Display:
     - Not correctly clearing text
-    - Improve performance by generating cleaning HTML markup
+    - Improve performance by generating cleaner HTML markup
     - MXP:
       - Fixed horizontal rulers display
       - Fixed selection highlighting of horizontal rulers
@@ -32,7 +35,12 @@
       - Fixed MXP send tags
     - Fixed HTML property to return correct HTML markup
     - Match whole unicode words on context or double click
-    - Throttle selection mouse moving to improve performance
+    - Debounce selection mouse moving to improve performance
+    - Fixed updating split view when scrolling
+    - Fixed split view selection performance issues
+    - Fixed scroll position on resize
+    - Fixed scrolling
+  - Fixed scroll lock scrolling to end when not in split screen mode
   - ANSI Parser: Fixed extra characters being added after : is detected, fixes raw ansi logging
 
 ## v0.6.4 2018-10-22